+++ packages/web3-eth-accounts/src/tx/index.ts
@@ -21,5 +21,5 @@ export { AccessListEIP2930Transaction } from './eip2930Transaction.js';
 export { Transaction } from './legacyTransaction.js';
 export { TransactionFactory } from './transactionFactory.js';
 export { BaseTransaction } from './baseTransaction.js';
-export * as txUtils from './utils';
+export * as txUtils from './utils.js';
 export * from './types.js';